+++ src/routes/BaseRoutes.tsx
@@ -1,24 +1,23 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 import App from '../App';
-import Login from '../pages/auth/Login';
 import Library from '../pages/common/Library';
 import Shelf from '../pages/common/Shelf';
 import AdminFeeds from '../pages/admin/AdminFeeds';
 import NotFound from '../pages/common/NotFound';
 import { NAVIGATION_PATHS } from '../utils/interfaces/general/general';
-import RequireAuth from './private/RequireAuth';
+import RequireAuth from './guards/AuthGuard';
 import useAuthContext from '../hooks/contexts/useAuthContext';
-import RequireAdmin from './private/RequireAdmin';
+import RequireAdmin from './guards/AdminGuard';
 import Viewer from '../pages/common/Viewer';
 import Home from '../pages/common/Home';
 import About from '../pages/common/About';
-import Loans from '../pages/common/Loans';
 import AdminEditEntry from '../pages/admin/entries/AdminEditEntry';
 import AdminEntries from '../pages/admin/entries/AdminEntries';
 import AdminCategories from '../pages/admin/AdminCategories';
 import Feeds from '../pages/common/Feeds';
 import AdminHome from '../pages/admin/AdminHome';
 import AdminAddEntry from '../pages/admin/entries/AdminAddEntry';
+import Auth from '../pages/auth/Auth';
 
 const BaseRoutes = () => {
   const { auth } = useAuthContext();
@@ -32,7 +31,7 @@ const BaseRoutes = () => {
         <Route
           path='login'
           element={
-            auth ? <Navigate to={NAVIGATION_PATHS.home} replace /> : <Login />
+            auth ? <Navigate to={NAVIGATION_PATHS.home} replace /> : <Auth />
           }
         />